#0: Moving ccl_op_fusion to a non-experimental folder.
