+++ ttnn/cpp/ttnn/operations/ccl/all_gather/device/all_gather_op.hpp
@@ -13,7 +13,7 @@
 #include "tt_metal/host_api.hpp"
 #include "ttnn/operations/ccl/ccl_host_datastructures.hpp"
 #include "ttnn/operations/ccl/ccl_common.hpp"
-#include "ttnn/operations/experimental/ccl/ccl_op_fusion.hpp"
+#include "ttnn/operations/ccl/ccl_op_fusion.hpp"
 
 
 #include "ttnn/run_operation.hpp"
